+++ src/Writer/EffectSchemaSchemaWriter.php
@@ -97,7 +97,9 @@ class EffectSchemaSchemaWriter implements SchemaWriter, Transformer
                 continue;
             }
 
-            $targetFilePath = $this->pathResolver->fqcnFilePath($info['fqcn']);
+            $targetFilePath = str_contains($info['fqcn'], '\\')
+                ? $this->pathResolver->fqcnFilePath($info['fqcn'])
+                : dirname($currentFilePath) . '/' . $info['alias'] . '.ts';
             $relativePath = $this->pathResolver->relativeImportPath(
                 $currentFilePath,
                 $targetFilePath,
@@ -200,6 +202,7 @@ class EffectSchemaSchemaWriter implements SchemaWriter, Transformer
                 && (
                     str_starts_with($type->fqcn, 'App\\')
                     || str_starts_with($type->fqcn, 'EffectSchemaGenerator\\')
+                    || !str_contains($type->fqcn, '\\')
                 )
             ) {
                 $referencedTypes[$type->fqcn] = [